requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-appconf==1.0.2
django-bower==5.2.0
django-bulma==0.5.0
django-compressor==2.2
django-compressor-toolkit==0.6.0
django-jsonify==0.3.0
django-markdown==0.8.4
django-markdownx==2.0.23
django-npm==1.0.0
django-simplemde==0.1.2
django-tastypie==0.14.1
django-yarn==0.0.1
idna==2.6
Markdown==2.6.11
oauthlib==2.0.7
Pillow==5.1.0
pkg-resources==0.0.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.4
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
urllib3==1.22
wadofstuff-django-serializers==1.1.0
